refactor(imgproc): de-duplicate imgproc module (#47)
Second de-duplication step of #47, following the pattern proven in #62.

- src/imgproc/imgproc.ts: replace the type-only stub with the REAL
  implementation moved verbatim from the monolith (grayscale, resample,
  box_blur_gray, gaussian_blur, hough_transform, pyrdown, scharr/sobel
  derivatives, compute_integral_image, equalize_histogram, canny,
  warp_perspective, warp_affine, skindetector). Only deliberate change:
  gaussian_blur instantiates the math module directly (import from
  ../math/math) instead of via the jsfeatNext.math static slot, removing
  an attach-order dependency on the aggregator.
- src/jsfeatNext.ts: shrinks by ~1050 lines; attaches imgproc from its
  module (jsfeatNext.imgproc = imgproc).
- Side effect: the latent trap in src/orb/rectify_patch.ts (it imports
  imgproc, which until now was the throwing stub) is healed — it resolves
  to the real implementation.

Verified behavior-preserving: tsc --noEmit clean; npm test 57/57 (14 of
those pin imgproc bit-for-bit vs original jsfeat); UMD build checked
(instanceof chain, static-constant inheritance, grayscale + gaussian_blur
smoke on the bundle).
